Move pseudo-bidomain code to package
